refactor: split test_clone.py (2019 lines) into four concern-focused files
Distributes 20 test classes across test_clone_core.py (~900 lines),
test_clone_push.py (~570 lines), test_clone_remote.py (~430 lines), and
test_clone_detect.py (~150 lines). Three shared fixtures migrate to a new
tests/workspace/conftest.py. TestPushToRemoteNonBare is decomposed by concern:
two clone_repo result-shape tests move to TestCloneRepoResultShape in core,
two push_to_remote tests stay in the push file.
