+++ tests/workspace/CLAUDE.md
@@ -9,7 +9,10 @@ Workspace cleanup, clone lifecycle, session skills, and worktree tests.
 | `__init__.py` | empty |
 | `conftest.py` | Shared fixtures for tests/workspace/ |
 | `test_cleanup.py` | L1 unit tests for workspace/cleanup.py — CleanupResult and directory deletion |
-| `test_clone.py` | Tests for autoskillit.workspace.clone module |
+| `test_clone_core.py` | Core clone_repo + remove_clone tests — setup, paths, error handling, origin contracts |
+| `test_clone_detect.py` | detect_source_dir, detect_branch, detect_uncommitted_changes, classify_remote_url |
+| `test_clone_push.py` | push_to_remote tests — E2E, mocked, protected branches, force-with-lease |
+| `test_clone_remote.py` | Remote resolution — probe helpers, URL resolution, stale-clone regression |
 | `test_clone_ci_contract.py` | Cross-boundary contract tests: clone isolation × CI/merge-queue resolution |
 | `test_clone_registry.py` | Tests for autoskillit.workspace.clone_registry module |
 | `test_clone_split.py` | Structural guard for clone test split |
@@ -27,4 +30,4 @@ Workspace cleanup, clone lifecycle, session skills, and worktree tests.
 
 ## Architecture Notes
 
-`conftest.py` provides shared fixtures for workspace tests. The `test_session_skills_*.py` files are split by concern across five files testing different aspects of the `session_skills` module.
+`conftest.py` provides shared fixtures for workspace tests. The `test_clone_*.py` files are split by concern from the original test_clone.py: core clone_repo behavior, push_to_remote, remote resolution/probing, and detect helpers. The `test_session_skills_*.py` files are split by concern across five files testing different aspects of the `session_skills` module.
